+++ packages/adapter-cloudflare-workers/index.js
@@ -6,18 +6,21 @@ import toml from '@iarna/toml';
 import { fileURLToPath } from 'url';
 
 /** @type {import('.')} */
-export default function () {
+export default function (options = {}) {
 	return {
 		name: '@sveltejs/adapter-cloudflare-workers',
 
 		async adapt(builder) {
-			const { site } = validate_config(builder);
+			const { site, build } = validate_config(builder);
 
 			// @ts-ignore
 			const { bucket } = site;
 
 			// @ts-ignore
 			const entrypoint = site['entry-point'] || 'workers-site';
+
+			// @ts-ignore
+			const main_path = build.upload.main;
 
 			const files = fileURLToPath(new URL('./files', import.meta.url).href);
 			const tmp = builder.getBuildDirectory('cloudflare-workers-tmp');
@@ -50,14 +53,20 @@ export default function () {
 			);
 
 			await esbuild.build({
-				entryPoints: [`${tmp}/entry.js`],
-				outfile: `${entrypoint}/index.js`,
-				bundle: true,
 				target: 'es2020',
-				platform: 'browser'
+				platform: 'browser',
+				...options,
+				entryPoints: [`${tmp}/entry.js`],
+				outfile: `${entrypoint}/${main_path}`,
+				bundle: true,
+				external: ['__STATIC_CONTENT_MANIFEST', ...(options?.external || [])],
+				format: 'esm'
 			});
 
-			writeFileSync(`${entrypoint}/package.json`, JSON.stringify({ main: 'index.js' }));
+			writeFileSync(
+				`${entrypoint}/package.json`,
+				JSON.stringify({ main: main_path, type: 'module' })
+			);
 
 			builder.log.minor('Copying assets...');
 			builder.writeClient(bucket);
@@ -86,6 +95,36 @@ function validate_config(builder) {
 			);
 		}
 
+		// @ts-ignore
+		if (!wrangler_config.build || !wrangler_config.build.upload) {
+			throw new Error(
+				'You must specify build.upload options in wrangler.toml. Consult https://github.com/sveltejs/kit/tree/master/packages/adapter-cloudflare-workers'
+			);
+		}
+
+		// @ts-ignore
+		if (wrangler_config.build.upload.format !== 'modules') {
+			throw new Error('build.upload.format in wrangler.toml must be "modules"');
+		}
+
+		// @ts-ignore
+		const main_file = wrangler_config.build?.upload?.main;
+		const main_file_ext = main_file?.split('.').slice(-1)[0];
+		if (main_file_ext && main_file_ext !== 'mjs') {
+			// @ts-ignore
+			const upload_rules = wrangler_config.build?.upload?.rules;
+			// @ts-ignore
+			const matching_rule = upload_rules?.find(({ globs }) =>
+				// @ts-ignore
+				globs.find((glob) => glob.endsWith(`*.${main_file_ext}`))
+			);
+			if (!matching_rule) {
+				throw new Error(
+					'To support a build.upload.main value not ending in .mjs, an upload rule must be added to build.upload.rules. Consult https://developers.cloudflare.com/workers/cli-wrangler/configuration/#build'
+				);
+			}
+		}
+
 		return wrangler_config;
 	}
 
@@ -103,6 +142,13 @@ function validate_config(builder) {
 		workers_dev = true
 		route = ""
 		zone_id = ""
+
+		[build]
+    command = ""
+
+		[build.upload]
+		format = "modules"
+		main = "./worker.mjs"
 
 		[site]
 		bucket = "./.cloudflare/assets"