+++ packages/create-svelte/templates/default/src/routes/todos/index.svelte
@@ -3,6 +3,16 @@
 	import { scale } from 'svelte/transition';
 	import { flip } from 'svelte/animate';
 
+	/**
+	 * @typedef {{
+	 *   uid: string;
+	 *   created_at: Date;
+	 *   text: string;
+	 *   done: boolean;
+	 *   pending_delete: boolean;
+	 * }} Todo
+	 */
+
 	type Todo = {
 		uid: string;
 		created_at: Date;
@@ -11,6 +21,7 @@
 		pending_delete: boolean;
 	};
 
+	/** @type {Todo[]} */
 	export let todos: Todo[];
 </script>