+++ packages/kit/src/core/sync/sync.js
@@ -2,10 +2,10 @@ import path from 'path';
 import create_manifest_data from './create_manifest_data/index.js';
 import { copy_assets } from './copy_assets.js';
 import { write_manifest } from './write_manifest.js';
+import { write_matchers } from './write_matchers.js';
 import { write_root } from './write_root.js';
 import { write_tsconfig } from './write_tsconfig.js';
 import { write_types } from './write_types.js';
-import { write_validators } from './write_validators.js';
 
 /** @param {import('types').ValidatedConfig} config */
 export function init(config) {
@@ -22,7 +22,7 @@ export function update(config) {
 
 	write_manifest(manifest_data, base, output);
 	write_root(manifest_data, output);
-	write_validators(manifest_data, output);
+	write_matchers(manifest_data, output);
 	write_types(config, manifest_data);
 
 	return { manifest_data };